+++ tests/Unit/Models/Article.php
@@ -10,7 +10,12 @@ class Article extends Model
     use CrudTrait;
 
     protected $table = 'articles';
-    protected $fillable = ['user_id', 'content', 'metas', 'tags', 'extras'];
+    protected $fillable = ['user_id', 'content', 'metas', 'tags', 'extras', 'cast_metas', 'cast_tags', 'cast_extras'];
+    protected $casts = [
+        'cast_metas' => 'object',
+        'cast_tags' => 'object',
+        'cast_extras' => 'object',
+    ];
 
     /**
      * Get the author for the article.